% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRevalue.R
\name{compare_revalue}
\alias{compare_revalue}
\title{Revalue the vector}
\usage{
compare_revalue(
  data,
  from = NULL,
  to = NULL,
  less = NULL,
  greater = NULL,
  equal = NULL,
  revalue = NULL
)
}
\arguments{
\item{data}{should be a vector, a matrix, or a dataframe}

\item{from}{character or factor vectors that need to be changed in data, this parameter is not used with the less, greater and equal parameter}

\item{to}{target vector, this parameter is used in conjunction with the from parameter, this parameter is not used with the less, greater and equal parameter}

\item{less}{if data are the numeric vectors, use this parameter as a comparison of less than objects}

\item{greater}{if data are the numeric vectors, use this parameter as a comparison of greater than objects}

\item{equal}{if data are the numeric vectors, use this parameter as a comparison of equality objects}

\item{revalue}{if data are the numeric vectors, use this parameter to specify the target vectors}
}
\value{

}
\description{
You can choose the same character or factor vectors in the data and change them to the target vectors.
If data are the numeric vectors can be changed to the target vectors by comparing sizes.
The target vectors can be character, factor, or numeric vectors.
}
\examples{
## Create a numeric vector
set.seed(123)
test_num <- runif(10, min=-1, max=2)
compare_revalue(data = test_num, less = 0.5, revalue = 1)

test_num1 <- c(0, 2,1,3,4,-0.5,-0.11,-0.25,0.5,1.5,2.5, 0.5)
compare_revalue(data = test_num1, great = 0.5, equal  = 0.5, revalue = 1)


## Construct a character vector
test_char <- rep("a", 10)
compare_revalue(test_char, from = "a", to = "A")


## Construct a factor vector
test_factor <- factor(test_char)
compare_revalue(test_char, from = "a", to = "A")


## Construct a matrix
set.seed(123)
test_matrix1 <- matrix(runif(1000000, min = -0.5, max = 1), nrow = 1000)
result1 <- compare_revalue(data = test_matrix1[1:100,1:100], less = 0.5, equal = 0.5, revalue = 0)
result1

set.seed(123)
test_matrix2 <- data.frame(a = sample(c("a","b","c"), 1000,replace = TRUE),
                          b = sample(c("A","B"), 1000, replace = TRUE),
                          c = sample(c("huawei","apple"), 1000, replace = TRUE),
                           d = sample(c(0, 1), 1000, replace = TRUE))
result2 <- compare_revalue(data = test_matrix2, from = "A", to = "C" )
result2

set.seed(123)
test_matrix3 <- data.frame(a = sample(c(1,2,3), 1000,replace = TRUE),
                          b = sample(c(5,6,7), 1000, replace = TRUE),
                          c = sample(c(-1,-2,-3), 1000, replace = TRUE),
                          d = sample(c(0, 1), 1000, replace = TRUE))
result3 <- compare_revalue(data = test_matrix3, equal = 0.5, revalue = 0)
result3

}
\author{
Haowei Ni <jylande@163.com >
}
