% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_from_3_cols.R
\name{date_from_3_cols}
\alias{date_from_3_cols}
\title{date_from_3_cols
takes a data frame, colname (string) to parse combines colname_year, _month, _date mutate newcol to paste to ymd drops colname_year, _month, _date}
\usage{
date_from_3_cols(df, date_col)
}
\arguments{
\item{df}{A data frame}

\item{date_col}{character string of name of a date column}
}
\value{

}
\description{
date_from_3_cols
takes a data frame, colname (string) to parse combines colname_year, _month, _date mutate newcol to paste to ymd drops colname_year, _month, _date
}
