% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_3_cols.R
\name{date_to_3_cols}
\alias{date_to_3_cols}
\title{date_to_3_cols
takes a dataframe, a (date) column
splits the column into separate year month date named with prefix
drop the original date column
This function assumes (without test) that date = yyyy/mm/dd}
\usage{
date_to_3_cols(df, col)
}
\arguments{
\item{df}{a dataframe}

\item{col}{a date column name string}
}
\value{

}
\description{
date_to_3_cols
takes a dataframe, a (date) column
splits the column into separate year month date named with prefix
drop the original date column
This function assumes (without test) that date = yyyy/mm/dd
}
