% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolory_tmp.R
\docType{data}
\name{kolory_nominalne_8}
\alias{kolory_nominalne_8}
\title{Palety kolorów zgodne z identyfikacją wizualną NASK}
\format{
Nazwane wektory tekstowe (character vector) zawierające kody HEX:
\describe{
  \item{kolory_nominalne_8}{paleta kolorów dla zmiennych nominalnych: 8 kolorów}
  \item{kolory_nominalne_12}{paleta kolorów dla zmiennych nominalnych: 12 kolorów}
  \item{kolory_unipolar_4}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 4 kolory}
  \item{kolory_unipolar_5}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 5 kolorów}
  \item{kolory_unipolar_7}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 7 kolorów}
  \item{kolory_unipolar_11}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 11 kolorów}
  \item{kolory_gradient}{kolor początkowy i końcowy gradientu dla zmiennych ciągłych}
  \item{kolory_bipolar_4}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 4 kolory}
  \item{kolory_bipolar_5}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 5 kolorów}
  \item{kolory_bipolar_7}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 7 kolorów}
  \item{kolory_bipolar_11}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 11 kolorów}
  \item{kolor_NA}{kolor dla wartości NA}
  \item{kolory_highlight}{dwa kolory analityczne dla wyróżniania elementów wykresu}
  \item{kolory_techniczne}{kolory techniczne dla elementów takich jak osie, tytuły etc.}
  
}
}
\source{
Własne opracowanie na podstawie księgi znaku NASK i podręcznika BAiB 1.2
}
\usage{
kolory_nominalne_8
}
\description{
Zbiór zdefiniowanych kolorów technicznych i marketingowych wykorzystywanych
w raportach oraz motywie \code{theme_nask}.
}
\examples{
TODO

}
\keyword{datasets}
