% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_nask_palettes.R
\name{show_nask_palettes}
\alias{show_nask_palettes}
\title{Wyświetl podgląd palet kolorów NASK}
\usage{
show_nask_palettes()
}
\value{
Obiekt ggplot2 prezentujący podgląd palet.
}
\description{
Funkcja generuje wykres (swatch plot) prezentujący wszystkie dostępne
w pakiecie palety kolorów, co ułatwia ich wybór do wizualizacji.
}
