% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_bloku_pytan.R
\name{tabela_bloku_pytan}
\alias{tabela_bloku_pytan}
\title{Tabela bloku pytań}
\usage{
tabela_bloku_pytan(
  .data = data,
  wzor_nazw_zmiennych = "",
  zmienne = NULL,
  tab_form = "wide"
)
}
\arguments{
\item{.data}{ramka danych}

\item{wzor_nazw_zmiennych}{(opcjonalnie) wspólny element nazw zmiennych
(string) pozwalający jednoznacznie zidentyfikować zmienne z bloku pytań}

\item{zmienne}{wektor nazw zmiennych}

\item{tab_form}{określa kształt tabeli: `wide` (szeroka) lub `long` (długa)}
}
\value{
ramke danych (tibble)
}
\description{
`tabela_bloku_pytan()` tworzy tabelę z rozkładem zmiennych pochodzących z
bloku pytań w kwestionariuszu, w których wykorzystano te same odpowiedzi
(wartości zmiennych sa takie same).
}
\examples{

tabela_bloku_pytan(dane,
                   zmienne = c('imd1_1_fct',
                               'imd1_2_fct',
                               'imd1_3_fct',
                               'imd1_4_fct',
                               'imd1_5_fct')
                   )

tabela_bloku_pytan(.data = dane, zmienne = paste0('p16_', 1:9, '_fct'))

tabela_bloku_pytan(.data = dane, wzor_nazw_zmiennych = 'p16_')

}
