% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolory.R
\docType{data}
\name{kolory_nask}
\alias{kolory_nask}
\alias{kolory_nominalne_8}
\alias{kolory_nominalne_12}
\alias{kolory_unipolar_4}
\alias{kolory_unipolar_5}
\alias{kolory_unipolar_7}
\alias{kolory_unipolar_11}
\alias{kolory_gradient}
\alias{kolory_bipolar_4}
\alias{kolory_bipolar_5}
\alias{kolory_bipolar_7}
\alias{kolory_bipolar_11}
\alias{kolory_diverging_4}
\alias{kolory_diverging_5}
\alias{kolory_diverging_7}
\alias{kolory_diverging_11}
\alias{kolor_NA}
\alias{kolory_highlight}
\alias{kolory_techniczne}
\title{Palety kolorów zgodne z identyfikacją wizualną NASK}
\format{
Nazwane wektory tekstowe (character vector) zawierające kody HEX:
\describe{
  \item{kolory_nominalne_8}{paleta kolorów dla zmiennych nominalnych: 8 kolorów}
  \item{kolory_nominalne_12}{paleta kolorów dla zmiennych nominalnych: 12 kolorów}
  \item{kolory_unipolar_4}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 4 kolory}
  \item{kolory_unipolar_5}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 5 kolorów}
  \item{kolory_unipolar_7}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 7 kolorów}
  \item{kolory_unipolar_11}{paleta kolorów dla zmiennych porządkowych-unipolarnych: 11 kolorów}
  \item{kolory_gradient}{kolor początkowy i końcowy gradientu dla zmiennych ciągłych}
  \item{kolory_bipolar_4}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 4 kolory}
  \item{kolory_bipolar_5}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 5 kolorów}
  \item{kolory_bipolar_7}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 7 kolorów}
  \item{kolory_bipolar_11}{paleta kolorów dla zmiennych porządkowych-bipolarnych: 11 kolorów}
  \item{kolor_NA}{kolor dla wartości NA}
  \item{kolory_highlight}{dwa kolory analityczne dla wyróżniania elementów wykresu}
  \item{kolory_techniczne}{kolory techniczne dla elementów takich jak osie, tytuły etc.}
  
}

An object of class \code{character} of length 8.

An object of class \code{character} of length 12.

An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 7.

An object of class \code{character} of length 11.

An object of class \code{character} of length 2.

An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 7.

An object of class \code{character} of length 11.

An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 7.

An object of class \code{character} of length 11.

An object of class \code{character} of length 1.

An object of class \code{list} of length 2.

An object of class \code{character} of length 3.
}
\source{
Własne opracowanie na podstawie księgi znaku NASK i podręcznika BAiB 1.2
}
\usage{
kolory_nominalne_8

kolory_nominalne_12

kolory_unipolar_4

kolory_unipolar_5

kolory_unipolar_7

kolory_unipolar_11

kolory_gradient

kolory_bipolar_4

kolory_bipolar_5

kolory_bipolar_7

kolory_bipolar_11

kolory_diverging_4

kolory_diverging_5

kolory_diverging_7

kolory_diverging_11

kolor_NA

kolory_highlight

kolory_techniczne
}
\description{
Zbiór zdefiniowanych kolorów technicznych i marketingowych wykorzystywanych
w raportach oraz motywie \code{theme_nask}.
}
\examples{
library(ggplot2)

# 1. Użycie palety nominalnej w barplocie
ggplot(mtcars, aes(x = factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_manual(values = kolory_nominalne_8) +
  theme_nask()

# 2. Użycie skali bipolarnej (np. dla korelacji lub skali Likerta)
df <- data.frame(x = 1:5, y = 1:5, z = c(-2, -1, 0, 1, 2))
ggplot(df, aes(x, y, color = z)) +
  geom_point(size = 10) +
  scale_color_gradientn(colors = kolory_bipolar_5) +
  theme_nask()

# 3. Dostęp do kolorów technicznych
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(color = kolory_techniczne['czarny']) +
  theme_nask()
  
}
\keyword{datasets}
