% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nask.R
\name{theme_nask}
\alias{theme_nask}
\title{Motyw graficzny zgodny ze standardem NASK}
\usage{
theme_nask(font_style = "thin")
}
\value{
Zwraca obiekt klasy \code{theme}, który można dodać do wykresu ggplot2.
}
\description{
Funkcja definiuje spójny wygląd wykresów ggplot2, ustawiając odpowiednie
czcionki (Archivo), kolory oraz rozmiary elementów tekstowych.
}
\details{
Motyw implementuje następujące wytyczne:
\itemize{
  \item \strong{Tytuł:} 16pt, Bold, czarny.
  \item \strong{Podtytuł:} 14pt, Bold, czarny.
  \item \strong{Osie/Etykiety:} 10pt, Regular, szary (#9D9D9D).
  \item \strong{Stopka:} 9pt, Regular, ciemnoszary (#757576).
}
}
\examples{
\ village {
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(title = "Tytuł wykresu", subtitle = "Podtytuł", caption = "Źródło: NASK") +
  theme_nask()
}
}
