% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_bloku_pytan.R
\name{wykres_bloku_pytan}
\alias{wykres_bloku_pytan}
\title{Wykres bloku pytań}
\usage{
wykres_bloku_pytan(
  .data = data,
  wzor_nazw_zmiennych = "",
  zmienne = NULL,
  weight = NULL,
  ordered = FALSE,
  rev_values = FALSE,
  sort_value = NULL,
  last_value = NULL,
  items_labels = NULL,
  items_labels_width = 60,
  values_text_cutoff = 5,
  digits = 1,
  bars_width = 0.8,
  fill_color_palette = "Greys",
  fill_color_direction = 1,
  fill_labels = NULL,
  legend_n_row = 2,
  legend.position = "top",
  legend_rev = FALSE,
  coord_flip = TRUE,
  title = "",
  subtitle = "",
  caption = ""
)
}
\arguments{
\item{.data}{ramka danych}

\item{wzor_nazw_zmiennych}{(opcjonalnie) wspólny element nazw zmiennych
(string) pozwalający jednoznacznie zidentyfikować zmienne z bloku pytań}

\item{zmienne}{wektor nazw zmiennych}

\item{ordered}{(opcjonalnie) TRUE/FALSE czy zmienne są typu `ordered`}

\item{rev_values}{TRUE/FALSE czy odwrócić kolejność wartości zmiennej na
wykresie}

\item{sort_value}{(string) wartość zmiennych, wg częstości której będa
sortowane zmienne na wykresie}

\item{last_value}{(string) zmienna, która ma być na ostatnim miejscu na
wykresie}

\item{items_labels}{wektor etykiet dla poszczególnych zmiennych
`c(nazwa_zmiennej = 'etykieta')`}

\item{items_labels_width}{liczba znaków zawijania etykiet zmienych}

\item{values_text_cutoff}{wartość etykiet wartości poniżej której etykiety
wartości nie sa wyświetlane na wykresie (domyślnie 5)}

\item{digits}{liczba miejsc po przecinku do zaokrąglenia etykiet wartości}

\item{bars_width}{szerokość słupków na wykresie. Domyślnie 0.8, mniejsze wartości - węższe słupki}

\item{fill_color_palette}{paleta dla wypełnień słupków}

\item{fill_color_direction}{kierunek układu kolorów na wykresie (1 lub -1 dla
odróconej kolejności)}

\item{fill_labels}{wektor etykiet dla poszczególnych kolorów wypełnienia}

\item{legend_n_row}{liczba wierszy dla elementów legendy "none", "left", "right", "bottom", "top", "inside". Wartość domyślna 'top'}

\item{legend.position}{pozycja legendy względem wykresu
#'@param legend_rev TRUE/FALSE czy odwrócić kolejność elementów w legendzie}

\item{coord_flip}{TRUE/FALSE czy odwrócić osie wykresu}

\item{title}{tytuł na wykresie}

\item{subtitle}{podtytuł na wykresie}

\item{caption}{informacja pod wykresem}
}
\value{
obiekt ggplot2
}
\description{
Funkcja worzy wykres przedstawiający rozkłady odpowiedzi na pytania z bloku
pytań kwestionariusza (np. pytania z oceną stwierdzeń). Zmienne przedstawiane
na wykresie powinny mieć takie same możliwe wartości
}
\examples{

p16_labels <- c(
p16_1_fct ='ChatGPT',
p16_2_fct ='Copilot',
p16_3_fct ='Gemini',
p16_4_fct ='Cloude',
p16_5_fct ='Perplexity',
p16_6_fct ='DeepSeek',
p16_7_fct ='PLLuM',
p16_8_fct ='Bielik',
p16_9_fct ='Inne'
)


wykres_bloku_pytan(.data = dane,
                  zmienne = paste0('p16_', 1:9, '_fct'),
                  rev_values = FALSE,
                  #fill_color_palette = niebieski_paleta[c(1, 9)]
                  legend_n_row = 1,
                  title = 'Narzędzia GenAI wykorzystywane w pracy przez pracowników #'administracji publicznej w Polsce',
                  subtitle = 'Z jakich narzędzi AI korzystał/a w ciągu ostatnich 6 miesięcy w #'pracy?',
                  caption = realizacja,
                  items_labels = p16_labels,
                  items_labels_width = 60,
                  sort_value = 'Zaznaczono',
                  last_value = 'p16_9_fct'
)

}
