% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_slupkowy.R
\name{wykres_slupkowy}
\alias{wykres_slupkowy}
\title{Wykres słupkowy}
\usage{
wykres_slupkowy(
  .data,
  column,
  cross_var = NULL,
  weight = NULL,
  fill_color = kolory_unipolar_5[5],
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  expand_top = 0.15,
  axix_x_labs_wrap = 20,
  co_flip = FALSE,
  rev_values = FALSE,
  highlighted_cat = NULL,
  highlight_fill = kolory_highlight$pomarancz,
  bars_width = 0.8,
  ...
)
}
\arguments{
\item{.data}{ramka danych}

\item{column}{nazwa zmiennej}

\item{cross_var}{TODO: nazwa zmiennej grupującej dla facet_wrap()}

\item{weight}{waga analityczna}

\item{fill_color}{kolor wypelnienia słupków}

\item{title}{tytuł na wykresie}

\item{subtitle}{podtytuł na wykresie}

\item{caption}{informacja pod wykresem}

\item{expand_top}{wartość określająca o ile zwiększyć oś y}

\item{axix_x_labs_wrap}{liczba znaków do zawijania etykiet wartości zmiennej na osi x}

\item{co_flip}{TRUE/FALSE czy odwrócic osie wykresu}

\item{rev_values}{TRUE/FALSE czy odwrócić kolejność wartości zmiennej na wykresie}

\item{highlighted_cat}{wartośc zmiennej do wyróżnienia osobnym kolorem}

\item{highlight_fill}{kolor wyróżnienia osobnej wartości (jeśli `highlighted_cat` != NULL)}

\item{...}{dodatkowe argumenty ggplot2}
}
\value{
a ggplot2 object
}
\description{
Wykres słupkowy
}
\examples{

mtcars$gear_fct <- as.factor(mtcars$gear)
wykres_slupkowy(.data = mtcars, column = gear_fct)

}
